refactor(#218): add index file in type folder
