+++ apps/frontend/src/features/drawing/utils/tableHelpers.ts
@@ -1,5 +1,5 @@
 import type { Node } from '@xyflow/react';
-import type { Table, Column, Constraint } from '@/types/erd.types';
+import type { Table, Column, Constraint } from '@/types';
 import type { TableData, ColumnType, ConstraintKind, Point } from '../types';
 
 type TableExtra = {
